ci/docker/docker-in-docker-image/publish-docker-image/requirements.txt
pyyaml==3.12
fire==0.1.0
testinfra==3.0.5
pytest==4.6.3
PyHamcrest==1.9.0
